+++ pkg/cli/start.go
@@ -35,7 +35,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/server"
-	"github.com/cockroachdb/cockroach/pkg/server/debug"
 	"github.com/cockroachdb/cockroach/pkg/server/status"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/storage"
@@ -205,66 +204,6 @@ func initMemProfile(ctx context.Context, dir string) {
 				}
 				gcProfiles(dir, memprof, maxSizePerProfile)
 			}()
-		}
-	}()
-}
-
-func initCPUProfile(ctx context.Context, dir string, st *cluster.Settings) {
-	const cpuprof = "cpuprof."
-	gcProfiles(dir, cpuprof, maxSizePerProfile)
-
-	cpuProfileInterval := envutil.EnvOrDefaultDuration("COCKROACH_CPUPROF_INTERVAL", -1)
-	if cpuProfileInterval <= 0 {
-		return
-	}
-	if min := time.Second; cpuProfileInterval < min {
-		log.Infof(ctx, "fixing excessively short cpu profiling interval: %s -> %s",
-			cpuProfileInterval, min)
-		cpuProfileInterval = min
-	}
-
-	go func() {
-		defer log.RecoverAndReportPanic(ctx, &serverCfg.Settings.SV)
-
-		ctx := context.Background()
-
-		t := time.NewTicker(cpuProfileInterval)
-		defer t.Stop()
-
-		var currentProfile *os.File
-		defer func() {
-			if currentProfile != nil {
-				pprof.StopCPUProfile()
-				currentProfile.Close()
-			}
-		}()
-
-		for {
-			// Grab a profile.
-			if err := debug.CPUProfileDo(st, cluster.CPUProfileDefault, func() error {
-				const format = "2006-01-02T15_04_05.999"
-
-				var buf bytes.Buffer
-				// Start the new profile. Write to a buffer so we can name the file only
-				// when we know the time at end of profile.
-				if err := pprof.StartCPUProfile(&buf); err != nil {
-					return err
-				}
-
-				<-t.C
-
-				pprof.StopCPUProfile()
-
-				suffix := timeutil.Now().Format(format)
-				if err := ioutil.WriteFile(filepath.Join(dir, cpuprof+suffix), buf.Bytes(), 0644); err != nil {
-					return err
-				}
-				gcProfiles(dir, cpuprof, maxSizePerProfile)
-				return nil
-			}); err != nil {
-				// Log errors, but continue. There's always next time.
-				log.Infof(ctx, "error during CPU profile: %s", err)
-			}
 		}
 	}()
 }
